split python paper interop support
